requirements.txt
python-chess>=1.0.0
ipython>=7.0.0
pygame>=2.1.0
